+++ apps/examples/dpp/philo.c
@@ -172,7 +172,7 @@ void philo_create(
     am_ao_create(
         &me->ao, (am_ao_fn)am_hsm_start, (am_ao_fn)am_hsm_dispatch, me
     );
-    am_hsm_create(&me->hsm, am_hsm_state_make(philo_init));
+    am_hsm_init(&me->hsm, am_hsm_state_make(philo_init));
 
     me->table = table;
     me->timer = timer;