+++ libs/hsm/tests/defer.c
@@ -128,7 +128,7 @@ static void defer_create(
     struct am_event_alloc* alloc
 ) {
     struct test_defer* me = &m_test_defer;
-    am_hsm_create(&me->hsm, am_hsm_state_make(defer_sinit));
+    am_hsm_init(&me->hsm, am_hsm_state_make(defer_sinit));
     me->log = log;
 
     /* setup HSM event queue */